Aggiornamento pre fine

+++ src/main/java/it/prova/triage/service/PazienteServiceImpl.java
@@ -1,5 +1,6 @@
 package it.prova.triage.service;
 
+import java.time.LocalDate;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,7 +8,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import it.prova.triage.model.Paziente;
+import it.prova.triage.model.StatoPaziente;
 import it.prova.triage.repository.paziente.PazienteRepository;
+import it.prova.triage.web.api.exception.NonRimovibileException;
+import it.prova.triage.web.api.exception.NotFoundException;
 
 @Service
 @Transactional
@@ -36,13 +40,53 @@ public class PazienteServiceImpl implements PazienteService{
 
 	@Override
 	public Paziente inserisciNuovo(Paziente pazienteInstance) {
+		pazienteInstance.setDataRegistrazione(LocalDate.now());
+		pazienteInstance.setStato(StatoPaziente.IN_ATTESA_VISITA);
 		return repository.save(pazienteInstance);
 		
 	}
 
 	@Override
 	public void rimuovi(Long idToRemove) {
+		Paziente pazienteDaControllare = repository.findById(idToRemove).orElse(null);
+		
+		if(pazienteDaControllare==null) {
+			throw new NotFoundException("Nessun paziente con questo id");
+		}
+		
+		if(!pazienteDaControllare.getStato().equals(StatoPaziente.DIMESSO)) {
+			throw new NonRimovibileException("Impossibile eliminare un paziente non dimesso");
+		}
 		repository.deleteById(idToRemove);
+	}
+
+	@Override
+	public void assegnaDottore(Long id, String codiceDottore) {
+		Paziente pazienteDaAssegnareDottore = repository.findById(id).orElse(null);
+		
+		if(pazienteDaAssegnareDottore==null) {
+			throw new NotFoundException("Nessun paziente con questo id non posso aggiungere un dottore");
+		}
+		
+		pazienteDaAssegnareDottore.setCodiceDottore(codiceDottore);
+		pazienteDaAssegnareDottore.setStato(StatoPaziente.IN_VISITA);
+		repository.save(pazienteDaAssegnareDottore);
+		
+		System.out.println(repository.findById(id).orElse(null).getCodiceDottore());
+	}
+
+	@Override
+	public void ricovera(Paziente paziente) {
+		paziente.setStato(StatoPaziente.RICOVERATO);
+		repository.save(paziente);
+	}
+
+	@Override
+	public void dimetti(Paziente paziente) {
+		paziente.setCodiceDottore(null);
+		paziente.setStato(StatoPaziente.DIMESSO);
+		repository.save(paziente);
+		
 	}
 
 }
